+++ mobile/android/android-components/components/browser/state/src/main/java/mozilla/components/browser/state/action/BrowserAction.kt
@@ -810,7 +810,7 @@ sealed class ContentAction : BrowserAction() {
     /**
      * Updates the [ContentState] of the given [sessionId] to indicate whether or not desktop mode is enabled.
      */
-    data class UpdateDesktopModeAction(val sessionId: String, val enabled: Boolean) : ContentAction()
+    data class UpdateTabDesktopMode(val sessionId: String, val enabled: Boolean) : ContentAction()
 
     /**
      * Updates the [AppIntentState] of the [ContentState] with the given [sessionId].
@@ -1921,4 +1921,19 @@ sealed class AppLifecycleAction : BrowserAction() {
      * The application has received an ON_PAUSE event.
      */
     object PauseAction : AppLifecycleAction()
+}
+
+/**
+ * [BrowserAction] implementations related to updating the application's default desktop mode setting.
+ */
+sealed class DefaultDesktopModeAction : BrowserAction() {
+    /**
+     * Toggles the global default for desktop browsing mode.
+     */
+    data object ToggleDesktopMode : DefaultDesktopModeAction()
+
+    /**
+     * Updates the global default for desktop browsing mode.
+     */
+    data class DesktopModeUpdated(val newValue: Boolean) : DefaultDesktopModeAction()
 }