+++ mobile/android/android-components/components/browser/state/src/main/java/mozilla/components/browser/state/engine/EngineObserver.kt
@@ -239,7 +239,7 @@ internal class EngineObserver(
 
     override fun onDesktopModeChange(enabled: Boolean) {
         store.dispatch(
-            ContentAction.UpdateDesktopModeAction(
+            ContentAction.UpdateTabDesktopMode(
                 tabId,
                 enabled,
             ),