+++ mobile/android/android-components/components/browser/state/src/main/java/mozilla/components/browser/state/reducer/BrowserStateReducer.kt
@@ -14,6 +14,7 @@ import mozilla.components.browser.state.action.CopyInternetResourceAction
 import mozilla.components.browser.state.action.CrashAction
 import mozilla.components.browser.state.action.CustomTabListAction
 import mozilla.components.browser.state.action.DebugAction
+import mozilla.components.browser.state.action.DefaultDesktopModeAction
 import mozilla.components.browser.state.action.DownloadAction
 import mozilla.components.browser.state.action.EngineAction
 import mozilla.components.browser.state.action.ExtensionsProcessAction
@@ -81,6 +82,7 @@ internal object BrowserStateReducer {
             is ExtensionsProcessAction -> ExtensionsProcessStateReducer.reduce(state, action)
             is AwesomeBarAction -> AwesomeBarStateReducer.reduce(state, action)
             is UpdateDistribution -> state.copy(distributionId = action.distributionId)
+            is DefaultDesktopModeAction -> DesktopModeReducer.reduce(state = state, action = action)
         }
     }
 }