+++ mobile/android/android-components/components/feature/tabs/src/main/java/mozilla/components/feature/tabs/TabsUseCases.kt
@@ -180,6 +180,7 @@ class TabsUseCases(
                 initialLoadFlags = flags,
                 initialAdditionalHeaders = additionalHeaders,
                 historyMetadata = historyMetadata,
+                desktopMode = store.state.desktopMode,
             )
 
             store.dispatch(TabListAction.AddTabAction(tab, select = selectTab))