+++ mobile/android/fenix/app/src/androidTest/java/org/mozilla/fenix/ui/SettingsSitePermissionsTest.kt
@@ -51,8 +51,8 @@ class SettingsSitePermissionsTest : TestSetup() {
         homeScreen {
         }.openThreeDotMenu {
         }.openSettings {
-        }.openSettingsSubMenuSitePermissions {
-            verifySitePermissionsToolbarTitle()
+        }.openSettingsSubMenuSiteSettings {
+            verifySiteSettingsToolbarTitle()
             verifyToolbarGoBackButton()
             verifySitePermissionOption("Autoplay", "Block audio only")
             verifySitePermissionOption("Camera", "Blocked by Android")
@@ -75,7 +75,7 @@ class SettingsSitePermissionsTest : TestSetup() {
         homeScreen {
         }.openThreeDotMenu {
         }.openSettings {
-        }.openSettingsSubMenuSitePermissions {
+        }.openSettingsSubMenuSiteSettings {
         }.openCamera {
             verifyBlockedByAndroidSection()
         }.goBack {
@@ -116,7 +116,7 @@ class SettingsSitePermissionsTest : TestSetup() {
         homeScreen {
         }.openThreeDotMenu {
         }.openSettings {
-        }.openSettingsSubMenuSitePermissions {
+        }.openSettingsSubMenuSiteSettings {
         }.openAutoPlay {
             verifySitePermissionsAutoPlaySubMenuItems()
             exitMenu()
@@ -181,7 +181,7 @@ class SettingsSitePermissionsTest : TestSetup() {
         homeScreen {
         }.openThreeDotMenu {
         }.openSettings {
-        }.openSettingsSubMenuSitePermissions {
+        }.openSettingsSubMenuSiteSettings {
         }.openAutoPlay {
             selectAutoplayOption("Allow audio and video")
             exitMenu()
@@ -216,7 +216,7 @@ class SettingsSitePermissionsTest : TestSetup() {
         homeScreen {
         }.openThreeDotMenu {
         }.openSettings {
-        }.openSettingsSubMenuSitePermissions {
+        }.openSettingsSubMenuSiteSettings {
         }.openAutoPlay {
             selectAutoplayOption("Allow audio and video")
             exitMenu()
@@ -243,7 +243,7 @@ class SettingsSitePermissionsTest : TestSetup() {
         homeScreen {
         }.openThreeDotMenu {
         }.openSettings {
-        }.openSettingsSubMenuSitePermissions {
+        }.openSettingsSubMenuSiteSettings {
         }.openAutoPlay {
             selectAutoplayOption("Block audio and video")
             exitMenu()
@@ -274,7 +274,7 @@ class SettingsSitePermissionsTest : TestSetup() {
         homeScreen {
         }.openThreeDotMenu {
         }.openSettings {
-        }.openSettingsSubMenuSitePermissions {
+        }.openSettingsSubMenuSiteSettings {
         }.openAutoPlay {
             selectAutoplayOption("Block audio and video")
             exitMenu()
@@ -310,7 +310,7 @@ class SettingsSitePermissionsTest : TestSetup() {
             navigationToolbar {
             }.openThreeDotMenu {
             }.openSettings {
-            }.openSettingsSubMenuSitePermissions {
+            }.openSettingsSubMenuSiteSettings {
             }.openCamera {
                 verifySitePermissionsCommonSubMenuItems()
                 selectPermissionSettingOption("Blocked")
@@ -336,7 +336,7 @@ class SettingsSitePermissionsTest : TestSetup() {
             navigationToolbar {
             }.openThreeDotMenu {
             }.openSettings {
-            }.openSettingsSubMenuSitePermissions {
+            }.openSettingsSubMenuSiteSettings {
             }.openMicrophone {
                 verifySitePermissionsCommonSubMenuItems()
                 selectPermissionSettingOption("Blocked")
@@ -361,7 +361,7 @@ class SettingsSitePermissionsTest : TestSetup() {
             navigationToolbar {
             }.openThreeDotMenu {
             }.openSettings {
-            }.openSettingsSubMenuSitePermissions {
+            }.openSettingsSubMenuSiteSettings {
             }.openLocation {
                 verifySitePermissionsCommonSubMenuItems()
                 selectPermissionSettingOption("Blocked")
@@ -386,7 +386,7 @@ class SettingsSitePermissionsTest : TestSetup() {
             navigationToolbar {
             }.openThreeDotMenu {
             }.openSettings {
-            }.openSettingsSubMenuSitePermissions {
+            }.openSettingsSubMenuSiteSettings {
             }.openNotification {
                 verifyNotificationSubMenuItems()
                 selectPermissionSettingOption("Blocked")
@@ -411,7 +411,7 @@ class SettingsSitePermissionsTest : TestSetup() {
             navigationToolbar {
             }.openThreeDotMenu {
             }.openSettings {
-            }.openSettingsSubMenuSitePermissions {
+            }.openSettingsSubMenuSiteSettings {
             }.openPersistentStorage {
                 verifySitePermissionsPersistentStorageSubMenuItems()
                 selectPermissionSettingOption("Blocked")
@@ -434,7 +434,7 @@ class SettingsSitePermissionsTest : TestSetup() {
             browserScreen {
             }.openThreeDotMenu {
             }.openSettings {
-            }.openSettingsSubMenuSitePermissions {
+            }.openSettingsSubMenuSiteSettings {
             }.openDRMControlledContent {
                 verifyDRMControlledContentSubMenuItems()
                 selectDRMControlledContentPermissionSettingOption("Blocked")
@@ -446,7 +446,7 @@ class SettingsSitePermissionsTest : TestSetup() {
                 verifyPageContent("DRM-controlled content not allowed")
             }.openThreeDotMenu {
             }.openSettings {
-            }.openSettingsSubMenuSitePermissions {
+            }.openSettingsSubMenuSiteSettings {
             }.openDRMControlledContent {
                 selectDRMControlledContentPermissionSettingOption("Allowed")
                 exitMenu()
@@ -472,7 +472,7 @@ class SettingsSitePermissionsTest : TestSetup() {
         }.clickPagePermissionButton(true) {
         }.openThreeDotMenu {
         }.openSettings {
-        }.openSettingsSubMenuSitePermissions {
+        }.openSettingsSubMenuSiteSettings {
         }.openExceptions {
             verifyExceptionCreated(permissionsTestPageHost, true)
             clickClearPermissionsOnAllSites()
@@ -494,7 +494,7 @@ class SettingsSitePermissionsTest : TestSetup() {
         }.clickPagePermissionButton(true) {
         }.openThreeDotMenu {
         }.openSettings {
-        }.openSettingsSubMenuSitePermissions {
+        }.openSettingsSubMenuSiteSettings {
         }.openExceptions {
             verifyExceptionCreated(permissionsTestPageHost, true)
             openSiteExceptionsDetails(permissionsTestPageHost)
@@ -517,7 +517,7 @@ class SettingsSitePermissionsTest : TestSetup() {
         }.clickPagePermissionButton(true) {
         }.openThreeDotMenu {
         }.openSettings {
-        }.openSettingsSubMenuSitePermissions {
+        }.openSettingsSubMenuSiteSettings {
         }.openExceptions {
             verifyExceptionCreated(permissionsTestPageHost, true)
             openSiteExceptionsDetails(permissionsTestPageHost)